FIN Google Forms service

+++ TTMMBot/Modules/Admin/AdminOwnerModule.cs
@@ -1,20 +1,25 @@
-﻿using System;
+﻿using Discord;
+using Discord.Commands;
+using System;
+using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
-using Discord;
-using Discord.Commands;
 using System.IO.Compression;
-using TTMMBot.Helpers;
+using System.Linq;
+using System.Net.Http;
+using System.Threading.Tasks;
 using TTMMBot.Modules.Interfaces;
 
 namespace TTMMBot.Modules.Admin
 {
     public partial class AdminModule : MMBotModule, IAdminModule
     {
+        private string _backupDir = Path.Combine(".", "backup");
+        private string _export = "dataexport.zip";
+        private string _import = "dataexport.zip";
+
         [Command("ExportDb")]
-        [Summary("Exports db data as json")]
+        [Summary("Exports db data as json files in zip archive")]
         [RequireOwner()]
         public async Task ExportDb()
         {
@@ -22,24 +27,66 @@ namespace TTMMBot.Modules.Admin
             {
                 var json = await _jsonService.ExportDBToJson();
 
-                var ex = await Task.Run(() => 
+                var ex = await Task.Run(async () =>
                 {
-                    var backupDir = Path.Combine(".", "backup");
-                    var export = "dataexport.zip";
+                    Directory.CreateDirectory(_backupDir);
 
-                    Directory.CreateDirectory(backupDir);
-                   
-                    foreach(var entry in json)
-                        File.WriteAllText(Path.Combine(backupDir, $"{entry.Key}.json"), entry.Value);
+                    foreach (var entry in json)
+                        await File.WriteAllTextAsync(Path.Combine(_backupDir, $"{entry.Key}.json"), entry.Value);
 
-                    ZipFile.CreateFromDirectory(backupDir, export);
+                    ZipFile.CreateFromDirectory(_backupDir, _export);
 
-                    Directory.Delete(backupDir, true);
-                    return export;
+                    Directory.Delete(_backupDir, true);
+                    return _export;
                 });
 
                 await Context.Channel.SendFileAsync(ex, "dbExport");
                 File.Delete(ex);
+            }
+            catch (Exception e)
+            {
+                await ReplyAsync($"{e.Message}");
+            }
+        }
+
+        [Command("ImportDb")]
+        [Summary("Imports the db with help of json files in zip archive")]
+        [RequireOwner()]
+        public async Task ImportDb()
+        {
+            try
+            {
+                await ReplyAsync("Starting db import this can take a while...");
+
+                var csvFile = Context.Message.Attachments.FirstOrDefault();
+                var myWebClient = _clientFactory.CreateClient();
+
+                if (csvFile != null)
+                {
+                    var zip = await myWebClient.GetAsync(csvFile.Url);
+                    var zipByte = await zip.Content.ReadAsByteArrayAsync();
+
+                    await File.WriteAllBytesAsync(_import, zipByte);
+
+                    var dict = await Task.Run(async () =>
+                    {
+                        Directory.CreateDirectory(_backupDir);
+
+                        ZipFile.ExtractToDirectory(_import, _backupDir);
+
+                        var dict = new Dictionary<string, string>();
+
+                        foreach (var entry in Directory.GetFiles(_backupDir))
+                            dict.Add(Path.GetFileNameWithoutExtension(entry), await File.ReadAllTextAsync(entry));
+
+                        Directory.Delete(_backupDir, true);
+                        return dict;
+                    });
+
+                    var result = await _jsonService.ImportJsonToDB(dict);
+                    File.Delete(_import);
+                    await ReplyAsync(result ? "db import completed!" : "error in db import!");
+                }
             }
             catch (Exception e)
             {
@@ -54,7 +101,7 @@ namespace TTMMBot.Modules.Admin
         [RequireOwner]
         public async Task Restart(bool saveRestart = true)
         {
-            if(saveRestart)
+            if (saveRestart)
             {
                 var r = await _databaseService?.AddRestart();
                 r.Guild = Context.Guild.Id;
@@ -64,7 +111,7 @@ namespace TTMMBot.Modules.Admin
 
             Process.Start(AppDomain.CurrentDomain.FriendlyName);
             await ReplyAsync($"Bot service is restarting...");
-            
+
             Environment.Exit(0);
         }
 
@@ -79,50 +126,24 @@ namespace TTMMBot.Modules.Admin
             await Restart(false);
         });
 
-        [Command("Show")]
-        [Summary("Show Global settings")]
-        [RequireOwner]
-        public async Task Show()
-        {
-            var gs = (await _databaseService.LoadGuildSettingsAsync());
-            var e = gs.GetEmbedPropertiesWithValues();
-            await ReplyAsync("", false, e as Embed);
-        }
-
-        [Command("Set")]
-        [Summary("Set Global settings")]
-        [RequireOwner]
-        public async Task Set(string propertyName, [Remainder] string value)
-        {
-            var gs = (await _databaseService.LoadGuildSettingsAsync());
-            var r = gs.ChangeProperty(propertyName, value);
-            await _databaseService.SaveDataAsync();
-            await ReplyAsync(r);
-        }
-
-        [Command("Get")]
-        [Summary("Get Global setting")]
-        [RequireOwner]
-        public async Task Get(string propertyName)
-        {
-            var gs = await _databaseService.LoadGuildSettingsAsync();
-            var r = gs.GetProperty(propertyName);
-            await ReplyAsync(r);
-        }
-
         [Command("AddChannelToUrlScan")]
         [Summary("Adds channel to url scan list")]
         [RequireOwner]
-        public async Task AddChannelToUrlScan(IGuildChannel channel)
+        public async Task AddChannelToUrlScan(IGuildChannel channel, IGuildChannel qChannel)
         {
-            await _commandHandler.AddChannelToGoogleFormsWatchList(channel);
+            _commandHandler.AddChannelToGoogleFormsWatchList(channel, qChannel);
 
-            var dbChannel = await _databaseService.CreateChannelAsync();
-            dbChannel.GuildId = channel.GuildId;
-            dbChannel.TextChannelId = channel.Id;
-            await _databaseService.SaveDataAsync();
+            var ch = await _databaseService.LoadChannelsAsync();
 
-            await ReplyAsync($"Successfully added {channel.Name} to UrlScanList.");
+            if(!ch.Any(c => c.GuildId == channel.GuildId && c.TextChannelId == channel.Id))
+            {
+                var dbChannel = await _databaseService.CreateChannelAsync();
+                dbChannel.GuildId = channel.GuildId;
+                dbChannel.TextChannelId = channel.Id;
+                dbChannel.AnswerTextChannelId = qChannel.Id;
+                await _databaseService.SaveDataAsync();
+            }
+            await ReplyAsync($"Successfully added {channel.Name} to UrlScanList and {qChannel.Name} for any questions!");
         }
 
         [Command("RemoveChannelFromUrlScan")]
@@ -130,10 +151,12 @@ namespace TTMMBot.Modules.Admin
         [RequireOwner]
         public async Task RemoveChannelFromUrlScan(IGuildChannel channel)
         {
-            await _commandHandler.RemoveChannelFromGoogleFormsWatchList(channel);
+            _commandHandler.RemoveChannelFromGoogleFormsWatchList(channel);
 
             var c = (await _databaseService.LoadChannelsAsync()).FirstOrDefault(x => x.GuildId == channel.GuildId && x.TextChannelId == channel.Id);
-            _databaseService.DeleteChannel(c);
+
+            if(c != null)
+                _databaseService.DeleteChannel(c);
 
             await ReplyAsync($"Successfully removed {channel.Name} to UrlScanList.");
         }
